Move OriginalDataPoint/DataSources to components structure
