+++ src/client/pages/OriginalDataPoint/OriginalDataPoint.scss
@@ -96,6 +96,7 @@
 
 .odp__data-source-input-column {
   width: 75%;
+  white-space: initial;
 
   .jodit-container {
     border: unset !important;
