+++ src/client/pages/OriginalDataPoint/components/DataSources/AdditionalComments/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AdditionalComments'
