+++ src/client/pages/OriginalDataPoint/components/DataSources/MethodsUsed/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MethodsUsed'
