+++ src/client/pages/OriginalDataPoint/components/DataSources/References/index.ts
@@ -1,0 +1,1 @@
+export { default } from './References'
